refactor models to zio.http.model + move Header type one package up
